* updated ellipse intersections and sampling \n * corrected randomized behaviours \n * sampling restriction is now a double (radius of small aggregates ignored)
git-svn-id: svn+ssh://c4science.ch/source/amie/trunk@3286 64fb8e13-074e-4207-b257-2d34fb7690c6

+++ features/features.h
@@ -114,7 +114,7 @@ protected:
 
     std::set<long unsigned int> changedDofs ;
     std::vector<Vector> reportValues ; 
-    size_t samplingRestriction ;
+    double samplingRestriction ;
 
     std::vector<std::vector<double>> cachedVolumes ;
     std::vector<Point *> extraPoints ;
@@ -388,7 +388,7 @@ public:
         return layer2d.size() ;
     }
 
-    void setSamplingRestriction ( size_t sr ) {
+    void setSamplingRestriction ( double sr ) {
         samplingRestriction = sr ;
     }
